[#6]/refactor components 파일 구조 변경
