+++ landing/src/components/Footer.tsx
