+++ landing/src/components/Section1.tsx
