+++ landing/src/components/Section2.tsx
