+++ landing/src/components/Section3.tsx
