+++ landing/src/pages/LandingPage.tsx
@@ -1,9 +1,9 @@
-import Footer from "components/footer";
-import Header from "components/header";
-import Section1 from "components/section1";
-import Section2 from "components/section2";
-import Section3 from "components/section3";
-import Section4 from "components/section4";
+import Footer from "components/Footer";
+import Header from "components/Header";
+import Section1 from "components/Section1";
+import Section2 from "components/Section2";
+import Section3 from "components/Section3";
+import Section4 from "components/Section4";
 import React, { ReactElement } from "react";
 import styled from "styled-components";
